requirements.txt
Django==3.1.5
django-appconf==1.0.4
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.2
mysqlclient==2.0.3
Pillow==8.1.0
